created the di class and made dependencies
