+++ app/src/main/java/com/example/git_friends/ui/MainActivity.kt
@@ -7,7 +7,9 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import com.example.git_friends.R
 import com.example.git_friends.data.App
+import com.example.git_friends.data.userentityrepo.UserEntityRepo
 import com.example.git_friends.databinding.ActivityMainBinding
+import com.example.git_friends.domain.UserEntity
 import com.example.git_friends.ui.listuserfragment.ListUsersFragment
 import com.example.git_friends.ui.userprofilefragment.UserProfileFragment
 
@@ -15,8 +17,9 @@ class MainActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityMainBinding
     private val listUsersFragment by lazy { ListUsersFragment(manageFragment) }
+    private val userProfileFragment by lazy { UserProfileFragment() }
 
-    private val userProfileFragment: UserProfileFragment = UserProfileFragment()
+    private val userEntityRepo: UserEntityRepo by lazy { App.instance.di.getInstanceUserEntityRepo() }
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,9 +27,10 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        App.instance.generateTestListUser() /** тестовый репозиторий*/
+        generateTestListUser() /** тестовый репозиторий*/
 
         initFragmentManager(listUsersFragment)
+
     }
 
      private fun initFragmentManager(fragment: Fragment) {
@@ -45,6 +49,16 @@ class MainActivity : AppCompatActivity() {
 
         }
 
+    }
+
+    fun generateTestListUser(){
+        userEntityRepo.createUser(UserEntity(1,"kshalnov"))
+        userEntityRepo.createUser(UserEntity(2,"Harrier55"))
+        userEntityRepo.createUser(UserEntity(3,"kshalnov"))
+        userEntityRepo.createUser(UserEntity(4,"Rogoz208"))
+        userEntityRepo.createUser(UserEntity(5,"niqmarin"))
+        userEntityRepo.createUser(UserEntity(6,"niqmarin"))
+        userEntityRepo.createUser(UserEntity(7,"test log 7"))
     }
